requirements.txt
# Flask and Web Framework
Flask==3.1.2
Flask-CORS==6.0.1
Flask-SocketIO==5.5.1
python-socketio==5.15.0

# PyTorch with CUDA 12.1 support
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121

# YOLO and Computer Vision
ultralytics==8.3.233
opencv-python==4.12.0.88

# Data Processing
numpy==2.2.6
scipy==1.16.3
polars==1.35.2

# Utilities
requests==2.32.5
pyyaml==6.0.3
psutil==7.1.3
ruamel.yaml==0.18.5
bcrypt==4.1.2

# Visualization (optional, for debugging)
matplotlib==3.10.7

